# Conflicts:
#	Application/Application.csproj
#	Application/Interfaces/IRequestService.cs
#	Application/Mappings/AutoMapperProfile.cs
#	Application/Services/RequestService.cs
#	Application/Services/UserService.cs
#	Domain/Interfaces/IRequestRepository.cs
#	Infrastructure/Repositories/RequestRepository.cs
#	Presentation/Controllers/CompanyController.cs
#	Presentation/Controllers/EquipmentController.cs
#	Presentation/Controllers/RequestsController.cs
#	Presentation/Program.cs

+++ Application/Services/EquipmentPhotoRepository.cs
@@ -32,10 +32,7 @@ public class EquipmentPhotoRepository : IEquipmentPhotoRepository
     public async Task DeleteAsync(EquipmentPhotoDto photo)
     {
         var entity = await _domainRepository.GetByIdAsync(photo.Id);
-        if (entity != null)
-        {
-            await _domainRepository.DeleteAsync(entity);
-        }
+        if (entity != null) await _domainRepository.DeleteAsync(entity);
     }
 
     public async Task SaveChangesAsync()
